moved BookRowItem to the content class
